requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
boto3==1.24.5
botocore==1.27.5
cfnresponse==1.1.2
jmespath==1.0.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
